[2]  create a "later" function analogous to "lazy" function which can  delay the execution of the code in the "by lazy" code block (only "by" is the keyword in Kotlin, and "lazy" is just a higher-order function). The code in the by lazy block does not get executed until uriMatcher variable is called for the first time.
In the lazy function, a Delegate object is created and returned. When we call the p property, we are actually calling the getValue() method of the Delegate object, and then the getValue() method calls the Lambda expression passed in by the lazy function.  The code in the expression can then be executed, and the resulting value obtained after calling the p property is the return value of the last line of code in the Lambda expression.

Replace the lazy function directly with later function.  Verify that the lazy loading of the later function is in effect: create a p field, add a log in the later function block, reference p field in the button click event.

We shall find that when Activity gets started, the log in later function is not triggered. Only when we click the button will the log get printed out which means that the log statement gets triggered. When we click the button again, there is no new log because the code in the later block will be only triggered once.

We only demonstrate the mechanism of how lazy function lazily load, but there are more things in lazy function which we do not cover, such as multi-threading code, empty value handling, etc.

Use the lazy function provided by Kotlin in real world project!